#ESW-489: Shivani | Manish: WIP: move logic to currentObsMode
